Moved the DOM and the String serializers and parsers as static inner classes of their abstract parents to improve the readability of the package contents. Scanner classes now follow the naming pattern xxxxIWXXXMScanner instead of IWXXMxxxxScanner
