+++ src/main/java/fi/fmi/avi/converter/iwxxm/conf/IWXXMGenericBulletinConverter.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import org.w3c.dom.Document;
 
 import fi.fmi.avi.converter.AviMessageSpecificConverter;
-import fi.fmi.avi.converter.iwxxm.bulletin.generic.AbstractGenericBulletinIWXXMParser;
+import fi.fmi.avi.converter.iwxxm.bulletin.generic.GenericBulletinIWXXMParser;
 import fi.fmi.avi.model.bulletin.GenericMeteorologicalBulletin;
 
 @Configuration
@@ -14,12 +14,12 @@ public class IWXXMGenericBulletinConverter {
     // Parsers:
     @Bean
     public AviMessageSpecificConverter<Document, GenericMeteorologicalBulletin> genericBulletinIWXXMDOMParser() {
-        return new AbstractGenericBulletinIWXXMParser.DOM();
+        return new GenericBulletinIWXXMParser.AsDOM();
     }
 
     @Bean
     public AviMessageSpecificConverter<String, GenericMeteorologicalBulletin> genericBulletinIWXXMStringParser() {
-        return new AbstractGenericBulletinIWXXMParser.String();
+        return new GenericBulletinIWXXMParser.AsString();
     }
 
     // Serializers:
