--- src/main/java/fi/fmi/avi/converter/iwxxm/v21/airmet/AIRMETIWXXMStringSerializer.java
@@ -1,16 +1,0 @@
-package fi.fmi.avi.converter.iwxxm.v21.airmet;
-
-import org.w3c.dom.Document;
-
-import fi.fmi.avi.converter.ConversionException;
-import fi.fmi.avi.converter.ConversionHints;
-import icao.iwxxm21.AIRMETType;
-
-public class AIRMETIWXXMStringSerializer extends AbstractAIRMETIWXXMSerializer<String> {
-
-    @Override
-    protected String render(final AIRMETType airmet, final ConversionHints hints) throws ConversionException {
-        final Document result = renderXMLDocument(airmet, hints);
-        return renderDOMToString(result, hints);
-    }
-}
