--- src/main/java/fi/fmi/avi/converter/iwxxm/v21/airmet/AirmetIWXXMSerializer.java
@@ -1,4 +1,0 @@
-package fi.fmi.avi.converter.iwxxm.v21.airmet;
-
-public class AirmetIWXXMSerializer {
-}
