+++ src/main/java/fi/fmi/avi/converter/iwxxm/v21/metar/SPECIIWXXMParser.java
@@ -4,8 +4,12 @@ import java.util.Objects;
 
 import javax.xml.bind.JAXBElement;
 
+import org.w3c.dom.Document;
+
+import fi.fmi.avi.converter.ConversionException;
 import fi.fmi.avi.converter.ConversionHints;
 import fi.fmi.avi.converter.ConversionResult;
+import fi.fmi.avi.converter.iwxxm.IWXXMConverterBase;
 import fi.fmi.avi.converter.iwxxm.ReferredObjectRetrievalContext;
 import fi.fmi.avi.model.metar.SPECI;
 import fi.fmi.avi.model.metar.immutable.SPECIImpl;
@@ -14,7 +18,7 @@ import icao.iwxxm21.SPECIType;
 /**
  * Created by rinne on 01/08/2018.
  */
-public abstract class AbstractSPECIIWXXMParser<T> extends MeteorologicalAerodromeObservationReportIWXXMParserBase<T, SPECI, SPECIImpl.Builder> {
+public abstract class SPECIIWXXMParser<T> extends MeteorologicalAerodromeObservationReportIWXXMParserBase<T, SPECI, SPECIImpl.Builder> {
 
     @Override
     protected SPECIImpl.Builder getEmptyBuilder() {
@@ -43,6 +47,36 @@ public abstract class AbstractSPECIIWXXMParser<T> extends MeteorologicalAerodrom
             return retval.build();
         } else {
             return null;
+        }
+    }
+
+    public static class AsDOM extends SPECIIWXXMParser<Document> {
+
+        /**
+         * Returns the SPECI input message as A DOM Document.
+         *
+         * @param input the XML Document input as a String
+         * @return the input parsed as DOM
+         * @throws ConversionException if an exception occurs while converting input to DOM
+         */
+        @Override
+        protected Document parseAsDom(final Document input) throws ConversionException {
+            return input;
+        }
+    }
+
+    public static class AsString extends SPECIIWXXMParser<String> {
+
+        /**
+         * Returns the SPECI input message as A DOM Document.
+         *
+         * @param input the XML Document input as a String
+         * @return the input parsed as DOM
+         * @throws ConversionException if an exception occurs while converting input to DOM
+         */
+        @Override
+        protected Document parseAsDom(final String input) throws ConversionException {
+            return IWXXMConverterBase.parseStringToDOM(input);
         }
     }
 }
