--- src/main/java/fi/fmi/avi/converter/iwxxm/v21/sigmet/SigmetIWXXMSerializer.java
@@ -1,4 +1,0 @@
-package fi.fmi.avi.converter.iwxxm.v21.sigmet;
-
-public class SigmetIWXXMSerializer {
-}
