--- src/main/java/fi/fmi/avi/converter/iwxxm/v21/taf/TAFIWXXMJAXBSerializer.java
@@ -1,16 +1,0 @@
-package fi.fmi.avi.converter.iwxxm.v21.taf;
-
-import fi.fmi.avi.converter.ConversionHints;
-import icao.iwxxm21.TAFType;
-
-/**
- * Specialization of {@link AbstractTAFIWXXMSerializer} for generating DOM output.
- */
-public class TAFIWXXMJAXBSerializer extends AbstractTAFIWXXMSerializer<TAFType> {
-
-    @Override
-    protected TAFType render(final TAFType taf, final ConversionHints hints) {
-        return taf;
-    }
-
-}
