--- src/main/java/fi/fmi/avi/converter/iwxxm/v30/SpaceWeatherAdvisory/SpaceWeatherIWXXMStringParser.java
@@ -1,13 +1,0 @@
-package fi.fmi.avi.converter.iwxxm.v30.SpaceWeatherAdvisory;
-
-import org.w3c.dom.Document;
-
-import fi.fmi.avi.converter.ConversionException;
-import fi.fmi.avi.converter.iwxxm.IWXXMConverterBase;
-
-public class SpaceWeatherIWXXMStringParser extends AbstractSpaceWeatherIWXXMParser<String> {
-    @Override
-    protected Document parseAsDom(final String input) throws ConversionException {
-        return IWXXMConverterBase.parseStringToDOM(input);
-    }
-}
